@@ -16,19 +16,19 @@
   },
   "homepage": "https://github.com/zorn-v/nextcloud-social-login#readme",
   "dependencies": {
-    "@nextcloud/axios": "^1.6.0",
+    "@nextcloud/axios": "^2.3.0",
     "@nextcloud/dialogs": "^1.4.0",
     "@nextcloud/password-confirmation": "^1.0.1",
-    "@nextcloud/router": "^2.0.0",
+    "@nextcloud/router": "^2.0.1",
     "css-loader": "^5.2.7",
-    "mini-css-extract-plugin": "^2.3.0",
-    "sass": "^1.40.1",
-    "sass-loader": "^12.1.0",
-    "vue": "^2.6.14",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.14",
-    "webpack": "^5.52.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-remove-empty-scripts": "^0.7.1"
+    "mini-css-extract-plugin": "^2.7.3",
+    "sass": "^1.59.3",
+    "sass-loader": "^12.6.0",
+    "style-loader": "^3.3.2",
+    "vue": "^3.2.41",
+    "vue-loader": "^17.0.1",
+    "webpack": "^5.76.1",
+    "webpack-cli": "^4.10.0",
+    "webpack-remove-empty-scripts": "^0.7.3"
   }
 }
