@@ -1,7 +1,10 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
+import { appName } from './common'
 import SettingsView from './components/Settings.vue'
 
-new Vue({
-  el: '#sociallogin_settings_app',
-  render: h => h(SettingsView),
-})
+createApp(SettingsView)
+  .use((app) => {
+    app.config.globalProperties.t = t
+    app.config.globalProperties.appName = appName
+  })
+  .mount('#sociallogin_settings_app')
