@@ -2,7 +2,7 @@
 
 // autoload_classmap.php @generated by Composer
 
-$vendorDir = dirname(dirname(__FILE__));
+$vendorDir = dirname(__DIR__);
 $baseDir = $vendorDir;
 
 return array(
@@ -60,6 +60,7 @@ return array(
     'Hybridauth\\Provider\\GitLab' => $vendorDir . '/hybridauth/hybridauth/src/Provider/GitLab.php',
     'Hybridauth\\Provider\\Google' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Google.php',
     'Hybridauth\\Provider\\Instagram' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Instagram.php',
+    'Hybridauth\\Provider\\Keycloak' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Keycloak.php',
     'Hybridauth\\Provider\\LinkedIn' => $vendorDir . '/hybridauth/hybridauth/src/Provider/LinkedIn.php',
     'Hybridauth\\Provider\\Mailru' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Mailru.php',
     'Hybridauth\\Provider\\Medium' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Medium.php',
@@ -71,6 +72,7 @@ return array(
     'Hybridauth\\Provider\\Paypal' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Paypal.php',
     'Hybridauth\\Provider\\PaypalOpenID' => $vendorDir . '/hybridauth/hybridauth/src/Provider/PaypalOpenID.php',
     'Hybridauth\\Provider\\Pinterest' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Pinterest.php',
+    'Hybridauth\\Provider\\PlexTv' => $vendorDir . '/custom/Hybridauth/Provider/PlexTv.php',
     'Hybridauth\\Provider\\QQ' => $vendorDir . '/hybridauth/hybridauth/src/Provider/QQ.php',
     'Hybridauth\\Provider\\Reddit' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Reddit.php',
     'Hybridauth\\Provider\\Slack' => $vendorDir . '/hybridauth/hybridauth/src/Provider/Slack.php',
