@@ -2,9 +2,10 @@
 
 // autoload_psr4.php @generated by Composer
 
-$vendorDir = dirname(dirname(__FILE__));
+$vendorDir = dirname(__DIR__);
 $baseDir = $vendorDir;
 
 return array(
     'Hybridauth\\' => array($vendorDir . '/hybridauth/hybridauth/src'),
+    '' => array($vendorDir . '/custom'),
 );
