@@ -1,31 +1,31 @@
 <?php return array(
     'root' => array(
+        'name' => 'sociallogin/3rdparty',
         'pretty_version' => 'dev-master',
         'version' => 'dev-master',
+        'reference' => '2f6432da951789064ad535419e8b7cc9dcbcd3e7',
         'type' => 'library',
         'install_path' => __DIR__ . '/../',
         'aliases' => array(),
-        'reference' => '6245af84c457c21c759447d50320e6c059178031',
-        'name' => 'sociallogin/3rdparty',
         'dev' => true,
     ),
     'versions' => array(
         'hybridauth/hybridauth' => array(
-            'pretty_version' => '3.7.1',
-            'version' => '3.7.1.0',
+            'pretty_version' => 'v3.8.1',
+            'version' => '3.8.1.0',
+            'reference' => 'cda2ddac102c29872164677fe92900bfef2eb15c',
             'type' => 'library',
             'install_path' => __DIR__ . '/../hybridauth/hybridauth',
             'aliases' => array(),
-            'reference' => '0ec1b8cb8b230aae739422c5e42bc5733b3a7316',
             'dev_requirement' => false,
         ),
         'sociallogin/3rdparty' => array(
             'pretty_version' => 'dev-master',
             'version' => 'dev-master',
+            'reference' => '2f6432da951789064ad535419e8b7cc9dcbcd3e7',
             'type' => 'library',
             'install_path' => __DIR__ . '/../',
             'aliases' => array(),
-            'reference' => '6245af84c457c21c759447d50320e6c059178031',
             'dev_requirement' => false,
         ),
     ),
