@@ -106,13 +106,12 @@ final class Parser
      * needs to be improved
      *
      * @param $birthday
-     * @param $seperator
      *
      * @return array
      */
-    public function parseBirthday($birthday, $seperator)
+    public function parseBirthday($birthday)
     {
-        $birthday = date_parse($birthday);
+        $birthday = date_parse((string) $birthday);
 
         return [$birthday['year'], $birthday['month'], $birthday['day']];
     }
