@@ -7,8 +7,9 @@
 
 namespace Hybridauth\Provider;
 
-use Hybridauth\Exception\InvalidArgumentException;
-use Hybridauth\Exception\UnexpectedApiResponseException;
+use Composer\InstalledVersions;
+use Exception;
+use Firebase\JWT\ExpiredException;
 use Hybridauth\Exception\InvalidApplicationCredentialsException;
 use Hybridauth\Exception\UnexpectedValueException;
 
@@ -19,8 +20,8 @@ use Hybridauth\User;
 use phpseclib\Crypt\RSA;
 use phpseclib\Math\BigInteger;
 
-use \Firebase\JWT\JWT;
-use \Firebase\JWT\JWK;
+use Firebase\JWT\JWT;
+use Firebase\JWT\Key;
 
 /**
  * Apple OAuth2 provider adapter.
@@ -100,6 +101,13 @@ class Apple extends OAuth2
     {
         parent::initialize();
         $this->AuthorizeUrlParameters['response_mode'] = 'form_post';
+
+        if ($this->isRefreshTokenAvailable()) {
+            $this->tokenRefreshParameters += [
+                'client_id' => $this->clientId,
+                'client_secret' => $this->clientSecret,
+            ];
+        }
     }
 
     /**
@@ -110,7 +118,7 @@ class Apple extends OAuth2
         $keys = $this->config->get('keys');
         $keys['secret'] = $this->getSecret();
         $this->config->set('keys', $keys);
-        return parent::configure();
+        parent::configure();
     }
 
     /**
@@ -170,7 +178,7 @@ class Apple extends OAuth2
             // validate the token signature and get the payload
             $publicKeys = $this->apiRequest('keys');
 
-            \Firebase\JWT\JWT::$leeway = 120;
+            JWT::$leeway = 120;
 
             $error = false;
             $payload = null;
@@ -188,18 +196,20 @@ class Apple extends OAuth2
                     );
                     $pem = $rsa->getPublicKey();
 
-                    $payload = JWT::decode($id_token, $pem, ['RS256']);
+                    $payload = ($this->getJwtVersion() < '6.2') ?
+                        JWT::decode($id_token, $pem, ['RS256']) :
+                        JWT::decode($id_token, new Key($pem, 'RS256'));
                     break;
-                } catch (\Exception $e) {
+                } catch (Exception $e) {
                     $error = $e->getMessage();
-                    if ($e instanceof \Firebase\JWT\ExpiredException) {
+                    if ($e instanceof ExpiredException) {
                         break;
                     }
                 }
             }
 
             if ($error && !$payload) {
-                throw new \Exception($error);
+                throw new Exception($error);
             }
         }
 
@@ -285,5 +295,10 @@ class Apple extends OAuth2
         $secret = JWT::encode($data, $key_content, 'ES256', $key_id);
 
         return $secret;
+    }
+
+    private function getJwtVersion()
+    {
+        return InstalledVersions::getVersion('firebase/php-jwt');
     }
 }
