@@ -82,7 +82,7 @@ class Spotify extends OAuth2
      */
     protected function fetchBirthday(User\Profile $userProfile, $birthday)
     {
-        $result = (new Data\Parser())->parseBirthday($birthday, '-');
+        $result = (new Data\Parser())->parseBirthday($birthday);
 
         $userProfile->birthDay = (int)$result[0];
         $userProfile->birthMonth = (int)$result[1];
