@@ -82,9 +82,11 @@ class WeChat extends OAuth2
         unset($this->tokenExchangeParameters['client_id']);
         unset($this->tokenExchangeParameters['client_secret']);
 
-        $this->tokenRefreshParameters += [
-            'appid' => $this->clientId
-        ];
+        if ($this->isRefreshTokenAvailable()) {
+            $this->tokenRefreshParameters += [
+                'appid' => $this->clientId,
+            ];
+        }
 
         $this->apiRequestParameters = [
             'appid' => $this->clientId,
