@@ -3,16 +3,29 @@ document.addEventListener('DOMContentLoaded', function () {
   if (!tgData) {
     return
   }
-  var login = tgData.dataset.login
+  var botId = tgData.dataset.botId
   var redirectUrl = tgData.dataset.redirectUrl
-
-  var mainEl = document.getElementsByTagName('main')[0]
-  if (mainEl && (document.querySelector('form[name="login"]') || document.querySelector('.section.sociallogin-connect'))) {
-    var script = document.createElement('script')
-    script.src = 'https://telegram.org/js/telegram-widget.js?9'
-    script.dataset.size = 'large'
-    script.dataset.telegramLogin = login
-    script.dataset.authUrl = redirectUrl
-    mainEl.appendChild(script)
+  var tgBtn = document.querySelector('#alternative-logins .telegram')
+  if (tgBtn) {
+    var tgWidget = document.createElement('script')
+    tgWidget.src = 'https://telegram.org/js/telegram-widget.js?21'
+    document.head.appendChild(tgWidget)
+    tgBtn.onclick = function (e) {
+      e.preventDefault()
+      Telegram.Login.auth({bot_id: botId}, function (data) {
+        if (data) {
+          try {
+            var redirectObj = new URL(redirectUrl)
+            new URLSearchParams(data).forEach((v, k) => {
+              redirectObj.searchParams.set(k, v)
+            })
+            location = redirectObj.href
+          } catch (err) {
+            location = redirectUrl + '?' + new URLSearchParams(data)
+          }
+          document.body.style.display = 'none'
+        }
+      })
+    }
   }
 })
