@@ -1,55 +1,85 @@
 OC.L10N.register(
     "sociallogin",
     {
-        "Disable auto create new users": "Automatisches Anlegen neuer Benutzer deaktivieren",
-        "Create users with disabled account": "Benutzer mit einem deaktivierten Konto erstellen",
-        "Allow users to connect social logins with their account": "Benutzern erlauben, externe Konten mit Ihrem Konto zu verknüpfen",
-        "Prevent creating an account if the email address exists in another account": "Anlegen eines Kontos verhindern, wenn die Email-Adresse bereits von einem anderen Konto verwendet wird",
-        "Update user profile every login": "Benutzerprofil nach jeder Anmeldung aktualisieren",
-        "Do not prune not available user groups on login": "Nicht vorhandene Benutzergruppen bei der Anmeldung nicht wieder löschen",
-        "Automatically create groups if they do not exists": "Nicht existierende Gruppen automatisch erstellen",
-        "Restrict login for users without mapped groups": "Verhindern, dass sich Benutzer ohne gemappte Gruppe anmelden können",
-        "Restrict login for users without assigned groups": "Verhindern, dass sich Benutzer ohne zugeordnete Gruppe anmelden können",
-        "Disable notify admins about new users": "Admins nicht über neu angelegte Benutzer benachrichtigen",
-        "Hide default login": "Standard-Login ausblenden",
-
-        "None": "Keine",
-        "Save": "Speichern",
-
-        "Custom OpenID Connect": "Benutzerdefinierte OpenID Connect Anbindung",
-        "Custom OAuth2": "Benutzerdefinierte OAuth2 Anbindung",
-        "Internal name": "Interner Name",
-        "Title": "Anzeigename",
-        "API Base URL": "API Basis-URL",
-        "Base url": "Basis-URL",
-        "Authorize url": "Authorize-URL",
-        "Authorize url (can be relative to base URL)": "Authorize-URL (kann relativ zur Basis-URL sein)",
-        "Identifier url": "Identifier-URL",
-        "Token url (can be relative to base URL)": "Token-URL (kann relativ zur Basis-URL sein)",
-        "Token url": "Token-URL",
-        "User info URL (optional)": "Userinfo-URL (optional)",
-        "Profile url": "Profile-URL",
-        "Profile url (can be relative to base URL)": "Profile-URL (kann relativ zur Basis-URL sein)",
-        "Logout URL (optional)": "Logout-URL (optional)",
-        "Client Id": "Client ID",
-        "Profile Fields (optional, comma-separated)": "Profilfelder (optional, kommasepariert)",
-        "Button style": "Stil der Schaltfläche",
-        "Default group": "Standardgruppe",
-        "Add group mapping": "Gruppen-Mapping hinzufügen",
-
-        "Custom OAuth1": "Benutzerdefinierte OAuth1 Anbindung",
-        "Custom Discourse": "Benutzerdefinierte Discourse Anbindung",
-
-        "App id": "App ID",
-
-        "Allow login only from specified domain": "Nur von der angegebenen Domäne",
-        "Allow login only for specified organizations": "Nur für Mitglieder der Organisationen",
-        "Allow login only for specified workspace": "Nur für Mitglieder des Workspace",
-        "Allow login only for specified guilds": "Nur für Mitglieder des Servers",
-
-        "Settings for social login successfully saved": "Die Einstellungen für Social Login wurden erfolgreich gespeichert",
-        "Do you really want to remove {providerTitle} provider ?": "Möchten Sie den Autorisierungsdienst {providerTitle} wirklich entfernen?",
-        "Some error occurred while saving settings": "Es trat ein Fehler beim Speichern der Einstellungen auf",
-        "Confirm remove": "Entfernen bestätigen"
-    },
-    "nplurals=2; plural=(n > 1);");
+    "Log in with username or email" : "Mit Benutzername oder E-Mail anmelden",
+    "Provider name cannot be empty" : "Der Anbietername darf nicht leer sein.",
+    "Duplicate provider name \"%s\"" : "Doppelter Anbietername \"%s\"",
+    "Invalid provider name \"%s\". Allowed characters \"0-9a-z_.@-\"" : "Ungültiger Anbietername \"%s\". Erlaubte Zeichen \"0-9a-z_.@-\"",
+    "Log in with %s" : "Anmelden mit %s",
+    "Unknown %s provider: \"%s\"" : "Unbekannter %s-Anbieter: \"%s\"",
+    "Can not get identifier from provider" : "Kennung vom Anbieter kann nicht abgerufen werden.",
+    "Login from %s domain is not allowed for %s provider" : "Die Anmeldung von der %s-Domäne ist für den %s-Anbieter nicht zulässig.",
+    "Login is available only to members of the following GitHub organizations: %s" : "Die Anmeldung ist nur für Mitglieder der folgenden GitHub-Organisationen verfügbar: %s",
+    "Invalid setup for Bitbucket workspaces" : "Ungültige Einrichtung für Bitbucket-Arbeitsbereiche",
+    "Login is available only to members of the following Bitbucket workspaces: %s" : "Die Anmeldung ist nur für Mitglieder der folgenden Bitbucket-Workspaces verfügbar: %s",
+    "Login is available only to members of the following Discord guilds: %s" : "Die Anmeldung ist nur für Mitglieder der folgenden Discord-Gilden verfügbar: %s",
+    "Social login connect is disabled" : "Social-Login-Verbindung ist deaktiviert.",
+    "This account already connected" : "Dieses Konto ist bereits verbunden.",
+    "Users without assigned groups is not allowed to login, please contact support" : "Benutzer ohne zugewiesene Gruppen dürfen sich nicht anmelden, wende dich bitte an den Support.",
+    "Your user group is not allowed to login, please contact support" : "Deine Benutzergruppe darf sich nicht anmelden, bitte wende dich an den Support.",
+    "Auto creating new users is disabled" : "Das automatische Erstellen neuer Benutzer ist deaktiviert.",
+    "Email already registered" : "E-Mail-Adresse ist schon registriert.",
+    "New user created" : "Neuer Benutzer erstellt",
+    "User %s (%s) just created via social login" : "Benutzer %s (%s) gerade über Social Login erstellt",
+    "Social login" : "Social-Login",
+    "Social Login" : "Social-Login",
+    "Social login via OAuth or OpenID" : "Social Login über OAuth oder OpenID",
+    "Save" : "Speichern",
+    "Button style" : "Knopf-Stil",
+    "None" : "Nicht kategorisierte Rezepte",
+    "Default group" : "Standardgruppe",
+    "App id" : "App-Id",
+    "Secret" : "Geheimer Schlüssel",
+    "Allow login only from specified domain" : "Anmeldung nur von der angegebenen Domain erlauben",
+    "Allow login only for specified organizations" : "Anmeldung nur für angegebene Organisationen zulassen",
+    "Allow login only for specified workspace" : "Anmeldung nur für den angegebenen Workspace zulassen",
+    "Allow login only for specified guilds" : "Anmeldung nur für bestimmte Gilden zulassen",
+    "Settings for social login successfully saved" : "Einstellungen für Social Login gespeichert",
+    "Some error occurred while saving settings" : "Es ist ein Fehler beim Speichern der Einstellungen aufgetreten.",
+    "Do you really want to remove {providerTitle} provider ?" : "Möchtest du wirklich den Anbieter {providerTitle} entfernen?",
+    "Confirm remove" : "Entfernen bestätigen",
+    "Add group mapping" : "Gruppenzuordnung hinzufügen",
+    "Disable auto create new users" : "Automatische Erstellung neuer Benutzer deaktivieren",
+    "Create users with disabled account" : "Benutzer mit deaktiviertem Konto erstellen",
+    "Allow users to connect social logins with their account" : "Erlaube Benutzern, soziale Logins mit ihrem Konto zu verbinden",
+    "Prevent creating an account if the email address exists in another account" : "Verbiete das Erstellen eines Kontos, wenn die E-Mail-Adresse in einem anderen Konto vorhanden ist.",
+    "Update user profile every login" : "Das Benutzerprofil bei jeder Anmeldung aktualisieren",
+    "Do not prune not available user groups on login" : "Nicht verfügbare Benutzergruppen bei der Anmeldung nicht löschen",
+    "Automatically create groups if they do not exists" : "Gruppen automatisch erstellen, wenn sie nicht vorhanden sind",
+    "Restrict login for users without mapped groups" : "Die Anmeldung für Benutzer ohne zugeordnete Gruppen beschränken",
+    "Restrict login for users without assigned groups" : "Anmeldung für Benutzer ohne zugewiesene Gruppen beschränken",
+    "Disable notify admins about new users" : "Administratoren über neue Benutzer benachrichtigen deaktivieren",
+    "Hide default login" : "Standardanmeldung ausblenden",
+    "Button text without prefix" : "Schaltflächentext ohne Präfix",
+    "Custom OpenID Connect" : "Benutzerdefiniertes OpenID-Connect",
+    "Internal name" : "Interner Name",
+    "Title" : "Titel",
+    "Authorize url" : "URL autorisieren",
+    "Token url" : "Token-URL",
+    "Display name claim (optional)" : "Anspruch auf Anzeigenamen (optional)",
+    "User info URL (optional)" : "Benutzerinfo-URL (optional)",
+    "Logout URL (optional)" : "Abmelde-URL (optional)",
+    "Client Id" : "Client-ID",
+    "Client Secret" : "Geheime Zeichenkette des Clients",
+    "Scope" : "Anwendungsbereich",
+    "Groups claim (optional)" : "Anspruch auf Gruppen (optional)",
+    "Identifier url" : "Identifikator-URL",
+    "Custom OAuth2" : "Benutzerdefiniertes OAuth2",
+    "API Base URL" : "API-Basis-URL",
+    "Authorize url (can be relative to base URL)" : "URL autorisieren (kann relativ zur Basis-URL sein)",
+    "Token url (can be relative to base URL)" : "Token-URL (kann relativ zur Basis-URL sein)",
+    "Profile url (can be relative to base URL)" : "Profil-URL (kann relativ zur Basis-URL sein)",
+    "Scope (optional)" : "Umfang (optional)",
+    "Profile Fields (optional, comma-separated)" : "Profilfelder (optional, durch Kommas getrennt)",
+    "Custom OAuth1" : "Benutzerdefiniertes OAuth1",
+    "Profile url" : "Profil-URL",
+    "Consumer key" : "Verbraucherschlüssel",
+    "Consumer Secret" : "Verbrauchergeheimnis",
+    "Custom Discourse" : "Benutzerdefinierter Discourse",
+    "Base url" : "Basis-URL",
+    "SSO Secret" : "SSO-Geheimnis",
+    "Disable password confirmation on settings change" : "Passwortbestätigung bei Einstellungsänderung deaktivieren",
+    "Social login connect" : "Soziale Anmeldung verbinden",
+    "Available providers" : "Verfügbare Anbieter"
+},
+"nplurals=2; plural=(n != 1);");
