@@ -1,16 +1,85 @@
 OC.L10N.register(
     "sociallogin",
     {
-        "Disable auto create new users": "Désactiver la création automatique de nouveaux utilisateurs",
-        "Allow users to connect social logins with their account": "Autoriser les utilisateurs existants à se connecter via social login",
-        "Prevent creating an account if the email address exists in another account": "Empêcher la création d'un compte si l'adresse e-mail existe déjà dans un autre compte",
-
-        "None": "Aucun",
-        "Save": "Sauvegarder",
-
-        "Settings for social login successfully saved": "Paramètres sauvegardés avec succès",
-        "Do you really want to remove {providerTitle} provider ?": "Voulez-vous vraiment supprimer fournisseur {providerTitle} ?",
-        "Some error occurred while saving settings": "Erreur lors de la sauvegarde des paramètres",
-        "Confirm remove": "Confirmer la suppression"
-    },
-    "nplurals=2; plural=(n > 1);");
+    "Log in with username or email" : "Se connecter avec un nom d'utilisateur ou une adresse e-mail",
+    "Provider name cannot be empty" : "Le nom du fournisseur ne peut pas être vide",
+    "Duplicate provider name \"%s\"" : "Duplicata du fournisseur \"%s\"",
+    "Invalid provider name \"%s\". Allowed characters \"0-9a-z_.@-\"" : "Nom du fournisseur \"%s\" invalide. Les caractères autorisés sont \"0-9a-z_.@-\".",
+    "Log in with %s" : "Se connecter avec %s",
+    "Unknown %s provider: \"%s\"" : "Fournisseur de %s inconnu : \"%s\"",
+    "Can not get identifier from provider" : "Impossible d'obtenir un identifiant depuis le fournisseur",
+    "Login from %s domain is not allowed for %s provider" : "La connexion depuis le domaine %s n'est pas autorisée pour le fournisseur %s",
+    "Login is available only to members of the following GitHub organizations: %s" : "La connexion est disponible uniquement pour les organisations GitHub suivantes: %s",
+    "Invalid setup for Bitbucket workspaces" : "Configuration invalide pour les espaces de travail Bitbucket",
+    "Login is available only to members of the following Bitbucket workspaces: %s" : "La connexion est disponible uniquement pour les espaces de travail Bitbucket suivants: %s",
+    "Login is available only to members of the following Discord guilds: %s" : "La connexion est disponible uniquement pour les guildes Discord suivantes: %s",
+    "Social login connect is disabled" : "Les connexions via Social sont désactivées",
+    "This account already connected" : "Ce compte est déjà lié",
+    "Users without assigned groups is not allowed to login, please contact support" : "Les utilisateurs sans groupes assignés ne sont pas autorisés à se connecter. Veuillez contacter le support.",
+    "Your user group is not allowed to login, please contact support" : "Votre groupe d'utilisateur n'est pas autorisé à se connecter. Veuillez contacter le support.",
+    "Auto creating new users is disabled" : "La création automatique de nouveaux utilisateurs est desactivée.",
+    "Email already registered" : "E-mail déjà enregistré",
+    "New user created" : "Nouvel utilisateur créé",
+    "User %s (%s) just created via social login" : "L'utilisateur %s (%s) vient juste d'être créé par Connexion via Social",
+    "Social login" : "Connexion via Social",
+    "Social Login" : "Connexion via Social",
+    "Social login via OAuth or OpenID" : "Connexion via Social avec OAuth ou OpenID",
+    "Save" : "Enregistrer",
+    "Button style" : "Style du bouton",
+    "None" : "Aucune",
+    "Default group" : "Groupe par défaut",
+    "App id" : "ID d'application",
+    "Secret" : "Caché",
+    "Allow login only from specified domain" : "Autoriser la connexion uniquement depuis le domaine spécifié",
+    "Allow login only for specified organizations" : "Autoriser la connexion uniquement pour l'organisation spécifiée",
+    "Allow login only for specified workspace" : "Autoriser la connexion uniquement pour l'espace de travail spécifié",
+    "Allow login only for specified guilds" : "Autoriser la connexion uniquement pour la guilde spécifiée",
+    "Settings for social login successfully saved" : "Paramètres de connexion via Social enregistrés avec succès",
+    "Some error occurred while saving settings" : "Des erreurs se sont produites lors de l'enregistrement des paramètres",
+    "Do you really want to remove {providerTitle} provider ?" : "Voulez-vous vraiment supprimer le fournisseur {providerTitle} ?",
+    "Confirm remove" : "Confirmer la suppression",
+    "Add group mapping" : "Ajouter une correspondance de groupe",
+    "Disable auto create new users" : "Désactivation de la création automatique d'utilisateurs",
+    "Create users with disabled account" : "Créer les utilisateurs avec des comptes désactivés",
+    "Allow users to connect social logins with their account" : "Autoriser les utilisateurs à se connecter à leur compte via Social",
+    "Prevent creating an account if the email address exists in another account" : "Empêche la création d'un utilisateur si son e-mail est déjà utilisée par un autre compte",
+    "Update user profile every login" : "Mettre à jour le profil utilisateur à chaque connexion",
+    "Do not prune not available user groups on login" : "Ne pas élaguer les groupes d'utilisateurs non disponibles à la connexion",
+    "Automatically create groups if they do not exists" : "Créer les groupes automatiquement s'ils n'existent pas",
+    "Restrict login for users without mapped groups" : "Restreindre la connexion aux utilisateurs sans groupes cartographiés ",
+    "Restrict login for users without assigned groups" : "Restreindre la connexion aux utilisateurs sans groupes assignés",
+    "Disable notify admins about new users" : "Désactiver les notifications vers les administrateurs lors de la création d'utilisateurs",
+    "Hide default login" : "Cacher la connexion par défaut",
+    "Button text without prefix" : "Texte du bouton sans préfixe",
+    "Custom OpenID Connect" : "Connexion OpenID personnalisée",
+    "Internal name" : "Nom interne",
+    "Title" : "Fonction",
+    "Authorize url" : "URL d'autorisation",
+    "Token url" : "URL du jeton",
+    "Display name claim (optional)" : "Réclamation du nom d'affichage (optionnel)",
+    "User info URL (optional)" : "URL des informations utilisateurs (optionnel)",
+    "Logout URL (optional)" : "URL après déconnexion (optionnelle)",
+    "Client Id" : "Id client",
+    "Client Secret" : "Secret du client",
+    "Scope" : "Portée",
+    "Groups claim (optional)" : "Réclamation des groupes (optionnel)",
+    "Identifier url" : "URL de l'identifiant",
+    "Custom OAuth2" : "OAuth2 personnalisé",
+    "API Base URL" : "URL de base de l'API",
+    "Authorize url (can be relative to base URL)" : "URL d'autorisation (peut être relatif à l'URL de base)",
+    "Token url (can be relative to base URL)" : "URL du jeton (peut être relatif à l'URL de base)",
+    "Profile url (can be relative to base URL)" : "URL du profile (peut être relatif à l'URL de base)",
+    "Scope (optional)" : "Portée (optionnel)",
+    "Profile Fields (optional, comma-separated)" : "Champs des profils (optionnel, séparés par des virgules)",
+    "Custom OAuth1" : "OAuth1 personnalisé",
+    "Profile url" : "URL du profil",
+    "Consumer key" : "Clef du client",
+    "Consumer Secret" : "Secret du client",
+    "Custom Discourse" : "Discours client",
+    "Base url" : "URL base",
+    "SSO Secret" : "Secret SSO",
+    "Disable password confirmation on settings change" : "Désactiver la demande du mot de passe lors du changement des paramètres",
+    "Social login connect" : "Connexion via les réseaux sociaux",
+    "Available providers" : "Fournisseurs disponibles"
+},
+"nplurals=3; plural=(n == 0 || n == 1) ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");
