@@ -1,16 +1,86 @@
 OC.L10N.register(
     "sociallogin",
     {
-        "Disable auto create new users": "Desativar criação automática de novos usuários",
-        "Allow users to connect social logins with their account": "Permitir que os usuários vinculem contas externas à sua conta",
-        "Prevent creating an account if the email address exists in another account": "Impedir a criação de uma conta se o email já existir em outra conta",
-
-        "None": "Nenhum",
-        "Save": "Salvar",
-
-        "Settings for social login successfully saved": "Configurações de login social salvas com sucesso",
-        "Do you really want to remove {providerTitle} provider ?": "Tem certeza de que deseja remover o provedor {providerTitle}?",
-        "Some error occurred while saving settings": "Ocorreu um erro ao salvar as configurações",
-        "Confirm remove": "Confirmar remoção"
-    },
-    "nplurals=2; plural=(n > 1);");
+    "Log in with username or email" : "Entre com nome de usuário ou e-mail",
+    "Provider name cannot be empty" : "Nome do provedor não pode ser vazio",
+    "Duplicate provider name \"%s\"" : "Nome de provedor \"1%s\" duplicado",
+    "Invalid provider name \"%s\". Allowed characters \"0-9a-z_.@-\"" : "Nome de provedor \"%s\" inválido.\nCaracteres permitidos \"0-9a-z_.@-\"",
+    "Log in with %s" : "Entre com %s",
+    "Unknown %s provider: \"%s\"" : "Provedor %s desconhecido: \"%s\"",
+    "Can not get identifier from provider" : "Não foi possível obter identificador do provedor",
+    "Login from %s domain is not allowed for %s provider" : "Login a partir do domínio %snão é permitido pelo provedor %s",
+    "Login is available only to members of the following GitHub organizations: %s" : "Login está disponível apenas para membros das organizações no GitHub: %s",
+    "Invalid setup for Bitbucket workspaces" : "Setup inválido para Bitbucket workspaces",
+    "Login is available only to members of the following Bitbucket workspaces: %s" : "Login está disponível apenas para membros dos workspaces Bitbucket: %s",
+    "Login is available only to members of the following Discord guilds: %s" : "Login está disponível apenas para membros das guildas no Discord: %s",
+    "Social login connect is disabled" : "Conectar por login social está dessabilitado",
+    "This account already connected" : "Esta conta já está conectada",
+    "Users without assigned groups is not allowed to login, please contact support" : "Usuários sem grupos não podem fazer login, por favor contacte o suporte",
+    "Your user group is not allowed to login, please contact support" : "Seu grupo de usuário não tem permissão para login, por favor contacte o suporte",
+    "Auto creating new users is disabled" : "Auto criação de usuários novos está desabilitado",
+    "Email already registered" : "E-mail já registrado",
+    "New user created" : "Novo usuário criado",
+    "User %s (%s) just created via social login" : "Usuario %s(%s) criado através de login social",
+    "Social login" : "Login Social",
+    "Social Login" : "Login Social",
+    "Social login via OAuth or OpenID" : "Login Social via OAuth ou OpenID",
+    "Save" : "Salvar",
+    "Button style" : "Estilo do botão",
+    "None" : "Nenhum",
+    "Default group" : "Grupo padrão",
+    "Add default provider" : "Adicionar provedor padrão",
+    "App id" : "ID do aplicativo",
+    "Secret" : "Secreto",
+    "Allow login only from specified domain" : "Permitir login apenas do domínio especificado",
+    "Allow login only for specified organizations" : "Permitir login apenas para organizações especificadas",
+    "Allow login only for specified workspace" : "Permitir login apenas para o espaço de trabalho especificado",
+    "Allow login only for specified guilds" : "Permitir login apenas para guildas especificadas",
+    "Settings for social login successfully saved" : "Configurações para login social salvas com sucesso",
+    "Some error occurred while saving settings" : "Ocorreu algum erro ao salvar as configurações",
+    "Do you really want to remove {providerTitle} provider ?" : "Deseja realmente remover o provedor {providerTitle}?",
+    "Confirm remove" : "Confirmar remover",
+    "Add group mapping" : "Adicionar mapeamento de grupo",
+    "Disable auto create new users" : "Desabilitar criação automática de novos usuários",
+    "Create users with disabled account" : "Criar usuários com conta desabilitada",
+    "Allow users to connect social logins with their account" : "Permitir que usuários conectem logins sociais às suas contas",
+    "Prevent creating an account if the email address exists in another account" : "Previne a criação de uma conta se o endereço de e-mail já existir em outra conta",
+    "Update user profile every login" : "Atualizar perfil do usuário a cada login",
+    "Do not prune not available user groups on login" : "Não remover grupos de usuários indisponíveis no login",
+    "Automatically create groups if they do not exists" : "Criar grupos automaticamente se não existir",
+    "Restrict login for users without mapped groups" : "Restringir login apenas para usuários sem grupo definido",
+    "Restrict login for users without assigned groups" : "Restringir login para logins sem grupo designado",
+    "Disable notify admins about new users" : "Desativar notificação para administração sobre novos usuários",
+    "Hide default login" : "Ocultar o login padrão",
+    "Button text without prefix" : "Texto no botão sem prefixo",
+    "Custom OpenID Connect" : "Conexão OpenID customizada",
+    "Internal name" : "Nome interno",
+    "Title" : "Título",
+    "Authorize url" : "Url de autorização",
+    "Token url" : "Url para token",
+    "Display name claim (optional)" : "Reivindicação de nome de exibição (opcional)",
+    "User info URL (optional)" : "URL de informações do usuário (opcional)",
+    "Logout URL (optional)" : "URL de saída (opcional)",
+    "Client Id" : "ID do Cliente",
+    "Client Secret" : "Segredo do client",
+    "Scope" : "Alcance",
+    "Groups claim (optional)" : "Reivindicação de grupos (opcional)",
+    "Identifier url" : "URL do identificador",
+    "Custom OAuth2" : "OAuth2 personalizado",
+    "API Base URL" : "URL base da API",
+    "Authorize url (can be relative to base URL)" : "URL de autorização (pode ser relativo ao URL base)",
+    "Token url (can be relative to base URL)" : "URL do token (pode ser relativo ao URL base)",
+    "Profile url (can be relative to base URL)" : "URL do perfil (pode ser relativo ao URL base)",
+    "Scope (optional)" : "Escopo (opcional)",
+    "Profile Fields (optional, comma-separated)" : "Campos de perfil (opcional, separados por vírgula)",
+    "Custom OAuth1" : "OAuth1 personalizado",
+    "Profile url" : "URL do perfil",
+    "Consumer key" : "Chave do consumidor",
+    "Consumer Secret" : "Consumidor Secreto",
+    "Custom Discourse" : "Discurso Personalizado",
+    "Base url" : "Base url",
+    "SSO Secret" : "SSO Secreto",
+    "Disable password confirmation on settings change" : "Desativar confirmação de senha na alteração de configurações",
+    "Social login connect" : "Conexão de login social",
+    "Available providers" : "Provedores disponíveis"
+},
+"nplurals=3; plural=(n == 0 || n == 1) ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");
