@@ -1,26 +1,47 @@
 OC.L10N.register(
     "sociallogin",
     {
-        "Disable auto create new users": "Отключить автоматическое создание новых пользователей",
-        "Create users with disabled account": "Создавать пользователей с отключенным аккаунтом",
-        "Allow users to connect social logins with their account": "Позволять пользователям подключать социальные логины к своим аккаунтам",
-        "Prevent creating an account if the email address exists in another account": "Не позволять создавать аккаунт, если пользователь с таким email адресом уже существует",
-        "Update user profile every login": "Обновлять профиль пользователя при каждом логине",
-        "Do not prune not available user groups on login": "Не вычищать недоступные пользователю группы при логине",
-        "Automatically create groups if they do not exists": "Автоматически создавать несуществующие группы",
-        "Restrict login for users without mapped groups": "Запрещать вход пользователям без ассоциированных групп",
-        "Restrict login for users without assigned groups": "Запрещать вход пользователям без установленных групп",
-        "Disable notify admins about new users": "Отключить оповещение администраторов о новых пользователях",
-
-        "None": "Отсутствует",
-        "Save": "Сохранить",
-        "Allow login only from specified domain": "Позволять вход только с указаного домена",
-        "Allow login only for specified organizations": "Позволять вход только для организаций",
-        "Allow login only for specified guilds": "Позволять вход только для гильдий",
-
-        "Settings for social login successfully saved": "Настройки social login успешно сохранены",
-        "Do you really want to remove {providerTitle} provider ?": "Вы действительно хотите удалить поставщика {providerTitle} ?",
-        "Some error occurred while saving settings": "Во время сохранения настроек произошла ошибка",
-        "Confirm remove": "Подтверждение удаления"
-    },
-    "nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);");
+    "Log in with username or email" : "Войти через имя пользователя или email",
+    "Provider name cannot be empty" : "Название поставщика не может быть пустым",
+    "Duplicate provider name \"%s\"" : "Дубликат названия поставщика \"%s\"",
+    "Invalid provider name \"%s\". Allowed characters \"0-9a-z_.@-\"" : "Некорректное название поставщика \"%s\". Допустимые символы \"0-9a-z_.@-\"",
+    "Log in with %s" : "Войти через %s",
+    "Unknown %s provider: \"%s\"" : "Неизвестный поставщик %s: \"%s\"",
+    "Login from %s domain is not allowed for %s provider" : "Вход с домена %s не разрешен для поставщика %s",
+    "Login is available only to members of the following GitHub organizations: %s" : "Вход разрешен только для членов следующих GitHub организаций: %s",
+    "Social login connect is disabled" : "Привязки социальных логинов отключены",
+    "This account already connected" : "Этот аккаунт уже подключен",
+    "Users without assigned groups is not allowed to login, please contact support" : "Вход пользователей без группы запрещен, пожалуйста свяжитесь с техподдержкой",
+    "Your user group is not allowed to login, please contact support" : "Вход для вашей группы пользователей запрещен, пожалуйста свяжитесь с техподдержкой",
+    "Auto creating new users is disabled" : "Автосоздание новых пользователей отключено",
+    "Email already registered" : "Пользователь с таким email уже зарегистрирован",
+    "New user created" : "Создан новый пользователь",
+    "User %s (%s) just created via social login" : "Пользователь %s (%s) создан через social login",
+    "Save" : "Сохранить",
+    "None" : "Отсутствует",
+    "Allow login only from specified domain" : "Позволять вход только с указаного домена",
+    "Allow login only for specified organizations" : "Позволять вход только для организаций",
+    "Allow login only for specified guilds" : "Позволять вход только для гильдий",
+    "Settings for social login successfully saved" : "Настройки social login успешно сохранены",
+    "Some error occurred while saving settings" : "Во время сохранения настроек произошла ошибка",
+    "Do you really want to remove {providerTitle} provider ?" : "Вы действительно хотите удалить поставщика {providerTitle} ?",
+    "Confirm remove" : "Подтверждение удаления",
+    "Disable auto create new users" : "Отключить автоматическое создание новых пользователей",
+    "Create users with disabled account" : "Создавать пользователей с отключенным аккаунтом",
+    "Allow users to connect social logins with their account" : "Позволять пользователям подключать социальные логины к своим аккаунтам",
+    "Prevent creating an account if the email address exists in another account" : "Не позволять создавать аккаунт, если пользователь с таким email адресом уже существует",
+    "Update user profile every login" : "Обновлять профиль пользователя при каждом логине",
+    "Do not prune not available user groups on login" : "Не вычищать недоступные пользователю группы при логине",
+    "Automatically create groups if they do not exists" : "Автоматически создавать несуществующие группы",
+    "Restrict login for users without mapped groups" : "Запрещать вход пользователям без ассоциированных групп",
+    "Restrict login for users without assigned groups" : "Запрещать вход пользователям без установленных групп",
+    "Disable notify admins about new users" : "Отключить оповещение администраторов о новых пользователях",
+    "Hide default login" : "Скрыть вход по умолчанию",
+    "Button text without prefix" : "Текст на кнопках без префикса",
+    "Internal name" : "Внутреннее имя",
+    "Title" : "Название",
+    "Disable password confirmation on settings change" : "Отключить подтверждение паролем изменения настроек",
+    "Social login connect" : "Привязки социальных логинов",
+    "Available providers" : "Доступные поставщики"
+},
+"nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);");
