@@ -1,29 +1,15 @@
 OC.L10N.register(
     "sociallogin",
     {
-        "Disable auto create new users": "取消自动创建使用者",
-        "Create users with disabled account": "默认创建禁用的用户",
-        "Allow users to connect social logins with their account": "允许使用者使用社交账号登录",
-        "Prevent creating an account if the email address exists in another account": "如果社交账号的邮箱已经注册过Nextcloud,则不自动创建新的使用者",
-        "Update user profile every login": "每次登陆自动更新用户信息",
-        "Do not prune not available user groups on login": "登陆时不清理不存在的用户组",
-        "Automatically create groups if they do not exists": "自动创造不存在的组",
-        "Restrict login for users without mapped groups": "限制没有分组的用户登陆",
-        "Disable notify admins about new users": "关闭新用户注册通知",
-
-        "None": "无",
-        "Save": "保存",
-        "Internal name": "内部名字",
-        "Title": "名字",
-        "Authorize url": "授权URL",
-        "User info URL (optional)": "用户信息URL(可选)",
-        "Default group": "默认用户组",
-        "Add group mapping": "添加绑定组",
-        "Custom OAuth2": "自定义 OAuth2",
-
-        "Settings for social login successfully saved": "保存成功",
-        "Do you really want to remove {providerTitle} provider ?": "您确定要刪除{providerTitle}吗？",
-        "Some error occurred while saving settings": "保存设置时发生错误",
-        "Confirm remove": "确认删除"
-    },
-    "nplurals=1; plural=0;");
+    "Log in with username or email" : "使用用户名或电子邮箱登录",
+    "Log in with %s" : "使用%s登录",
+    "Save" : "保存",
+    "None" : "无",
+    "Secret" : "Secret",
+    "Title" : "标题",
+    "Client Id" : "客户端ID",
+    "Client Secret" : "客户端密钥",
+    "Scope" : "适用范围",
+    "Consumer key" : "网站密钥"
+},
+"nplurals=1; plural=0;");
