@@ -1,16 +1,86 @@
 OC.L10N.register(
     "sociallogin",
     {
-        "Disable auto create new users": "關閉自動新增使用者",
-        "Allow users to connect social logins with their account": "允許使用者使用社交登入",
-        "Prevent creating an account if the email address exists in another account": "若登入帳號的電子郵件與已註冊用戶之相符，則暫停註冊帳戶",
-
-        "None": "無",
-        "Save": "儲存",
-
-        "Settings for social login successfully saved": "儲存成功",
-        "Do you really want to remove {providerTitle} provider ?": "你真的想刪除這個{providerTitle}項目嗎？",
-        "Some error occurred while saving settings": "儲存設定時發生了錯誤",
-        "Confirm remove": "確定刪除"
-    },
-    "nplurals=1; plural=0;");
+    "Log in with username or email" : "以使用者名稱或電子郵件登入",
+    "Provider name cannot be empty" : "提供者名稱不能為空",
+    "Duplicate provider name \"%s\"" : "重複的提供者名稱「%s」",
+    "Invalid provider name \"%s\". Allowed characters \"0-9a-z_.@-\"" : "無效的提供者名稱「%s」。允許字元為「0-9a-z_.@-」",
+    "Log in with %s" : "以 %s 登入",
+    "Unknown %s provider: \"%s\"" : "未知的 %s 提供者：「%s」",
+    "Can not get identifier from provider" : "無法取得來自提供者的識別字串",
+    "Login from %s domain is not allowed for %s provider" : "來自 %s 網域的登入不允許 %s 提供者",
+    "Login is available only to members of the following GitHub organizations: %s" : "登入僅適用於以下 GitHub 組織的成員：%s",
+    "Invalid setup for Bitbucket workspaces" : "無效的 Bitbucket 工作空間設定",
+    "Login is available only to members of the following Bitbucket workspaces: %s" : "登入僅適用於以下 Bitbucket 工作空間的成員：%s",
+    "Login is available only to members of the following Discord guilds: %s" : "登入僅適用於以下 Discord 公會的成員：%s",
+    "Social login connect is disabled" : "社群登入連結已停用",
+    "This account already connected" : "此帳號已連結",
+    "Users without assigned groups is not allowed to login, please contact support" : "未指派群組的使用者將不允許登入，請聯絡支援",
+    "Your user group is not allowed to login, please contact support" : "您的使用者群組並不允許登入，請聯絡支援",
+    "Auto creating new users is disabled" : "自動建立新使用者已停用",
+    "Email already registered" : "電子郵件已註冊",
+    "New user created" : "新使用者已建立",
+    "User %s (%s) just created via social login" : "剛透過社交登入建立的使用者 %s (%s)",
+    "Social login" : "社交登入",
+    "Social Login" : "社交登入",
+    "Social login via OAuth or OpenID" : "透過 OAuth 或 OpenID 進行社交登入",
+    "Save" : "儲存",
+    "Button style" : "按鈕樣式",
+    "None" : "無",
+    "Default group" : "預設群組",
+    "Add default provider" : "新增預設提供者",
+    "App id" : "應用程式 ID",
+    "Secret" : "密碼",
+    "Allow login only from specified domain" : "僅允許來自特定網域的登入",
+    "Allow login only for specified organizations" : "僅允許特定組織登入",
+    "Allow login only for specified workspace" : "僅允許特定工作空間登入",
+    "Allow login only for specified guilds" : "僅允許特定公會登入",
+    "Settings for social login successfully saved" : "社交登入的設定已儲存",
+    "Some error occurred while saving settings" : "儲存設定時發生一些錯誤",
+    "Do you really want to remove {providerTitle} provider ?" : "您真的想要移除 {providerTitle} 提供者嗎？",
+    "Confirm remove" : "確認移除",
+    "Add group mapping" : "新增群組映射",
+    "Disable auto create new users" : "停用自動建立新使用者",
+    "Create users with disabled account" : "建立已停用帳號的使用者",
+    "Allow users to connect social logins with their account" : "允許使用者將他們的帳號與社交登入連結",
+    "Prevent creating an account if the email address exists in another account" : "若電子郵件地址已存在於另一個帳號中，則阻止建立帳號",
+    "Update user profile every login" : "每次登入都更新使用者個人檔案",
+    "Do not prune not available user groups on login" : "不要在登入時精簡不可用的使用者群組",
+    "Automatically create groups if they do not exists" : "若群組不存在，則自動建立",
+    "Restrict login for users without mapped groups" : "限制沒有映射群組的使用者登入",
+    "Restrict login for users without assigned groups" : "限制未指派群組的使用者的登入",
+    "Disable notify admins about new users" : "停用對管理員關於新使用者的通知",
+    "Hide default login" : "隱藏預設登入",
+    "Button text without prefix" : "沒有前綴的按鈕文字",
+    "Custom OpenID Connect" : "自訂 OpenID 連結",
+    "Internal name" : "內部名稱",
+    "Title" : "標題",
+    "Authorize url" : "授權 URL",
+    "Token url" : "權杖 URL",
+    "Display name claim (optional)" : "顯示名稱聲明（選擇性）",
+    "User info URL (optional)" : "使用者資訊 URL（選擇性）",
+    "Logout URL (optional)" : "登出 URL（選擇性）",
+    "Client Id" : "客戶端 ID",
+    "Client Secret" : "客戶端密碼",
+    "Scope" : "範圍",
+    "Groups claim (optional)" : "群組聲明（選擇性）",
+    "Identifier url" : "識別字串 URL",
+    "Custom OAuth2" : "自訂 OAuth2",
+    "API Base URL" : "API 基礎 URL",
+    "Authorize url (can be relative to base URL)" : "授權 URL（可以相對於基礎 URL）",
+    "Token url (can be relative to base URL)" : "權杖 URL（可以相對於基礎 URL）",
+    "Profile url (can be relative to base URL)" : "個人檔案 URL（可以相對於基礎 URL）",
+    "Scope (optional)" : "範圍（選擇性）",
+    "Profile Fields (optional, comma-separated)" : "個人檔案欄位（選擇性，以逗號分隔）",
+    "Custom OAuth1" : "自訂 OAuth1",
+    "Profile url" : "個人檔案 URL",
+    "Consumer key" : "消費者金鑰",
+    "Consumer Secret" : "消費者密碼",
+    "Custom Discourse" : "自訂 Discourse",
+    "Base url" : "基礎 URL",
+    "SSO Secret" : "SSO 密碼",
+    "Disable password confirmation on settings change" : "停用設定變更時的密碼確認",
+    "Social login connect" : "社交登入連結",
+    "Available providers" : "可用的提供者"
+},
+"nplurals=1; plural=0;");
