@@ -14,11 +14,16 @@ module.exports = {
   output: {
     path: path.resolve(__dirname, 'js'),
   },
+  performance: {
+    maxAssetSize: 256000,
+    maxEntrypointSize: 256000,
+  },
+  stats: { modulesSpace: 0 },
   module: {
     rules: [
       {
         test: /\.css$/,
-        use: ['vue-style-loader', 'css-loader']
+        use: ['style-loader', 'css-loader']
       },
       {
         test: /\.scss$/,
